fix nextApi name-convention
